Update build and move project-loader back to lib
- export image and sounds in dist/static/assets
- rename main export (`scratch-gui.js`, not `index.js`)
- move project-loader-hoc back to lib (VM doesn’t load properly without it)
